+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ### Fixes
 - Fix confirm reset pin error text is not hidden by default
+- Show invalid json error when `NHID` number is empty or less than 14 digits   
 
 ## 2021-06-16-7826
 ### Internal